+++ src/main/resources/dev/luin/fs/db/migration/postgresql/V1.0.0__Initial.sql
@@ -34,6 +34,6 @@ CREATE TABLE file
 	end_date					TIMESTAMP				NULL,
 	user_id						INTEGER					NOT NULL,
 	length						BIGINT					NULL,
-	type							TINYINT					NULL,
+	type							SMALLINT				NULL,
 	FOREIGN KEY (user_id) REFERENCES fs_user(id)
 );
